Refactor /주석 제거 및 import 함수 정리
