+++ src/apis/index.ts
@@ -1,1 +1,3 @@
 export * from './user';
+export * from './products';
+export * from './user';
